# frontend/public/swagger.yml
openapi: 3.0.3
info:
  title: Breezy API
  description: API REST du réseau social Breezy
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Serveur local de développement
...
paths:
...
  /posts:
    get:
      tags: [Posts]
      summary: Liste des posts
      responses:
        '200':
          description: Liste de posts
    post:
      tags: [Posts]
      summary: Créer un post
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
      responses:
        '201':
          description: Post créé
  /posts/{id}:
    delete:
      tags: [Posts]
      summary: Supprimer un post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Supprimé
  /posts/{id}/comments:
    get:
      tags: [Comments]
      summary: Obtenir les commentaires d’un post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Liste de commentaires

    post:
      tags: [Comments]
      summary: Commenter un post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
      responses:
        '201':
          description: Commentaire publié
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PostInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        mediaUrl:
          type: string
...